buck2_downward_api

Trait DownwardApi

pub trait DownwardApi {
    // Required methods
    fn console<'life0, 'async_trait>(
        &'life0 self,
        level: Level,
        msg: String,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn log<'life0, 'async_trait>(
        &'life0 self,
        level: Level,
        msg: String,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn external<'life0, 'async_trait>(
        &'life0 self,
        data: HashMap<String, String>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

The API available to processes that Buck will need to handle

Required Methods

fn console<'life0, 'async_trait>( &'life0 self, level: Level, msg: String, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

indicates to print to the console at a specific log level

fn log<'life0, 'async_trait>( &'life0 self, level: Level, msg: String, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

indicates to log at a specified level TODO consider if we should have structured log instead of a String message

fn external<'life0, 'async_trait>( &'life0 self, data: HashMap<String, String>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

reports an externally consumable event containing some data that will be untouched by buck